requirements.txt
Flask==3.0.0
psycopg2-binary
flask-mail
pymongo